chore: added mobile read-only editor
